Update test formatting and make more dirs
