+++ backend/events/tests/test_event_str_methods.py
@@ -3,22 +3,20 @@
 Tests for the events app models.
 """
 
-# mypy: ignore-errors
 import pytest
 
 from events.factories import (
-    EventFactory,
     EventAttendeeFactory,
     EventAttendeeStatusFactory,
+    EventFactory,
     FormatFactory,
     RoleFactory,
 )
-from events.models import EventText
 
 pytestmark = pytest.mark.django_db
 
 
-def test_str_methods() -> None:
+def test_event_str_methods() -> None:
     """
     Test string representation methods for event-related models.
     """
@@ -33,18 +31,3 @@ def test_str_methods() -> None:
     assert str(event_attendee_status) == event_attendee_status.status_name
     assert str(_format) == _format.name
     assert str(role) == role.name
-
-
-def test_event_text_str_method() -> None:
-    """
-    Test string representation method for EventText model.
-    """
-    event = EventFactory.create()
-    event_text = EventText.objects.create(
-        event=event,
-        iso="en",
-        primary=True,
-        description="Test description",
-        get_involved="Get involved text"
-    )
-    assert str(event_text) == f"{event_text.event} - {event_text.iso}"
